Split Subtitles and AVPlayerViewController extension
